+++ src/main/java/com/pocketful/web/dto/payment/NewPaymentDTO.java
@@ -2,6 +2,7 @@ package com.pocketful.web.dto.payment;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -10,7 +11,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class NewPaymentDTO {
-    private float amount;
+    private BigDecimal amount;
     private String description;
     private Boolean payed;
     private Boolean isExpense;
